TP5-POE

+++ src/fr/diginamic/model/SupprimerQuestionService.java
@@ -2,12 +2,21 @@ package fr.diginamic.model;
 
 import java.util.Scanner;
 
+import fr.diginamic.exception.SupprimerQuestionException;
+
 public class SupprimerQuestionService extends MenuService{
 	
-	public void executeUC(Scanner scanner, QuestionDao dao) {
+	@Override
+	public void executeUC(Scanner scanner, QuestionDao dao) throws SupprimerQuestionException{
+		
 		System.out.println("Veuillez indiquer le numéro de la question à supprimer (entre 1 et "+dao.findAll().size()+")");
 		int indexSuppr = Integer.parseInt(scanner.nextLine());
+		if(indexSuppr>dao.findAll().size()) {
+			throw new SupprimerQuestionException("Le nombre doit être inférieur à "+dao.findAll().size());
+			
+		}
 		dao.delete(dao.findAll().get(indexSuppr-1));
+		
 	}
 
 }
